PowerVM: Do not overwrite the original drive boot entry
Instead, add the ReaR entry before the original one.  Otherwise, if ReaR
failed and the machine was returned to Beaker, it would successfuly boot
from the network but the installer WOULD NOT change the boot order to the
disk drive with the fresh RHEL installation.

Therefore, after that OpenFirmware would just boot loop as the network
installation files were already gone and the second disk drive that contained
ReaR was erased.

With this change, it would just skip the network, empty disk drive and
continue to boot the new RHEL installation.